examples/autogen/requirements.txt
pyautogen>=0.2.0
python-dotenv>=1.0.0
openai>=1.3.0
termcolor>=2.3.0
